backend/requirement.txt
Flask==3.1.3
flask-cors==4.0.0
mysql-connector-python==8.2.0
Werkzeug==3.1.6
google-generativeai==0.3.1
PyPDF2==3.0.1
python-dotenv==1.0.0    